refactor(favorites): add src, types, and tests folders
