+++ packages/favorites/package.json
@@ -4,6 +4,7 @@
   "author": "Kyle Gray <kyle.gray@availity.com>",
   "description": "Favorite Heart for favoriting items such as links/resources etc.",
   "main": "index.js",
+  "types": "index.d.ts",
   "keywords": [
     "react",
     "availity"
@@ -29,7 +30,6 @@
   "devDependencies": {
     "@availity/api-axios": "^5.5.14",
     "@availity/api-core": "^7.0.2",
-    "@availity/localstorage-core": "^3.0.0",
     "axios": "^0.21.1",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
@@ -38,8 +38,7 @@
   "peerDependencies": {
     "@availity/api-axios": "^5.4.0",
     "@availity/api-core": "^5.0.0",
-    "@availity/localstorage-core": "^2.6.1",
-    "axios": "^0.19.2",
+    "axios": "^0.21.1",
     "react": ">=16.8.0",
     "reactstrap": "^8.0.0 || ^7.1.0"
   }
