+++ packages/favorites/tests/maxFavorites.json
