+++ packages/favorites/types/FavoriteHeart.d.ts
@@ -4,6 +4,7 @@ interface FavoriteHeartProps {
   name: string;
   onChange?: (isFavorited: boolean, event: React.MouseEvent<any>) => void;
 }
-declare const FavoriteHeart: React.FunctionComponent<FavoriteHeartProps>;
+
+declare const FavoriteHeart: React.FC<FavoriteHeartProps>;
 
 export default FavoriteHeart;
